# docs/swagger.yaml
definitions:
  models.Todo:
    properties:
      created_at:
        type: string
      description:
        type: string
      deadline:
        type: string
      id:
        type: integer
      title:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /todos:
    get:
      description: get all todos
      operationId: get-todos
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Todo'
            type: array
      summary: Get all todos
    post:
      consumes:
      - application/json
      description: create a new todo
      operationId: create-todo
      parameters:
      - description: request body json
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Create a todo
  /todos/{id}:
    delete:
      description: delete a todo
      operationId: delete-todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Delete a todo
    get:
      description: get a todo
      operationId: get-todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Get a todo
    put:
      consumes:
      - application/json
      description: update a todo
      operationId: update-todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: integer
      - description: request body json
        in: body
        name: requestbody
        required: true
        schema:
          $ref: '#/definitions/models.Todo'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Todo'
      summary: Update a todo
swagger: "2.0"